requirements.txt
streamlit==1.28.2
pandas==1.5.3
plotly==5.18.0
folium==0.14.0
streamlit-folium==0.16.0
google-cloud-bigquery==3.13.0